Context menus are super pretty now

+++ trunk/src/rutebaga/controller/command/list/StatsListElementFactory.java
@@ -11,7 +11,7 @@ public class StatsListElementFactory implements ListElementFactory<StatValue> {
 				return null;
 			}
 			public String getLabel() {
-				return stat.getId().getName() + ": " + stat.getValue();
+				return stat.getId().getName() + ": " + (int) stat.getValue();
 			}
 		};
 	}
